Wrote thread2, added some semaphores
I added a ton of semaphores to fit directly into the bounded buffer paradigm we discussed at the meeting. I am not certain if adhering to that format so exactly is necessary or not. I also created a binary semaphore for locking freelist that I did not use, instead opting to follow the format used in thread1.